refactor: correct PR errors
