+++ src/services/resources/token/getAnonymous/index.ts
