+++ src/services/resources/token/index.ts
@@ -1,1 +1,1 @@
-export * from './getAnonymousToken'
+export * from './getAnonymous'
